+++ set_test.go
@@ -1,7 +1,6 @@
 package set
 
 import (
-	"fmt"
 	"sync"
 	"testing"
 
@@ -207,56 +206,4 @@ func TestConcurrent(t *testing.T) {
 			})
 		})
 	}
-}
-
-func TestAtomicMarkableReference(t *testing.T) {
-	t.Run("construction", func(t *testing.T) {
-		testCases := []struct {
-			mark bool
-			val  int
-		}{
-			{
-				mark: false,
-				val:  10,
-			},
-			{
-				mark: true,
-				val:  10,
-			},
-		}
-
-		for i, tc := range testCases {
-			tc := tc
-
-			t.Run(fmt.Sprint(i), func(t *testing.T) {
-				node := &nodeNonBlocking{val: tc.val}
-
-				amr := newAtomicMarkableReference(node, tc.mark)
-				amr.getNode()
-
-				require.Equal(t, node, amr.getNode())
-				require.Equal(t, tc.val, amr.getNode().val)
-				require.Equal(t, tc.mark, amr.getMark())
-
-				nodeRead, markRead := amr.getBoth()
-				require.Equal(t, node, nodeRead)
-				require.Equal(t, tc.val, nodeRead.val)
-				require.Equal(t, tc.mark, markRead)
-			})
-		}
-	})
-
-	t.Run("mutation", func(t *testing.T) {
-		node1 := &nodeNonBlocking{val: 1}
-		node2 := &nodeNonBlocking{val: 2}
-		mark1 := true
-		mark2 := false
-
-		amr := newAtomicMarkableReference(node1, mark1)
-
-		require.True(t, amr.compareAndSet(node1, node2, mark1, mark2))
-
-		require.Equal(t, node2, amr.getNode())
-		require.Equal(t, mark2, amr.getMark())
-	})
 }
